+++ lib/mongoid/indexable.rb
@@ -1,13 +1,18 @@
 # encoding: utf-8
+require "mongoid/indexable/specification"
 require "mongoid/indexable/validators/options"
 
 module Mongoid
+
+  # Encapsulates behaviour around defining indexes.
+  #
+  # @since 4.0.0
   module Indexable
     extend ActiveSupport::Concern
 
     included do
-      cattr_accessor :index_options
-      self.index_options = {}
+      cattr_accessor :index_specifications
+      self.index_specifications = []
     end
 
     module ClassMethods
@@ -21,13 +26,14 @@ module Mongoid
       #
       # @since 1.0.0
       def create_indexes
-        return unless index_options
-        index_options.each_pair do |spec, options|
+        return unless index_specifications
+        index_specifications.each do |spec|
+          key, options = spec.key, spec.options
           if database = options[:database]
             with(consistency: :strong, database: database).
-              collection.indexes.create(spec, options.except(:database))
+              collection.indexes.create(key, options.except(:database))
           else
-            with(consistency: :strong).collection.indexes.create(spec, options)
+            with(consistency: :strong).collection.indexes.create(key, options)
           end
         end and true
       end
@@ -62,20 +68,18 @@ module Mongoid
       #
       # @since 1.0.0
       def add_indexes
-        if hereditary? && !index_options[{ _type: 1 }]
-          index({ _type: 1 }, { unique: false, background: true })
+        if hereditary? && !index_keys.include?(_type: 1)
+          index({ _type: 1 }, unique: false, background: true)
         end
         true
       end
 
-      # Adds an index on the field specified. Options can be :unique => true or
-      # :unique => false. It will default to the latter.
+      # Adds an index definition for the provided single or compound keys.
       #
       # @example Create a basic index.
       #   class Person
       #     include Mongoid::Document
       #     field :name, type: String
-
       #     index({ name: 1 }, { background: true })
       #   end
       #
@@ -86,52 +90,24 @@ module Mongoid
       #
       # @since 1.0.0
       def index(spec, options = nil)
-        Validators::Options.validate(self, spec, options || {})
-        index_options[normalize_spec(spec)] = normalize_index_options(options)
+        index_specifications.push(Specification.new(self, spec, options))
+      end
+
+      # Get an index specification for the provided key.
+      #
+      # @example Get the index specification.
+      #   Model.index_specification(name: 1)
+      #
+      # @param [ Hash ] key The index key/direction pair.
+      #
+      # @return [ Specification ] The found specification.
+      #
+      # @since 4.0.0
+      def index_specification(key)
+        index_specifications.detect{ |spec| spec.fields == key.keys }
       end
 
       private
-
-      # Normalize the index options, if any are provided.
-      #
-      # @api private
-      #
-      # @example Normalize the index options.
-      #   Model.normalize_index_options(drop_dups: true)
-      #
-      # @param [ Hash ] options The index options.
-      #
-      # @return [ Hash ] The normalized options.
-      #
-      # @since 3.0.0
-      def normalize_index_options(options)
-        opts = options || {}
-        opts[:dropDups] = opts.delete(:drop_dups) if opts.has_key?(:drop_dups)
-        opts[:bucketSize] = opts.delete(:bucket_size) if opts.has_key?(:bucket_size)
-        if opts.has_key?(:expire_after_seconds)
-          opts[:expireAfterSeconds] = opts.delete(:expire_after_seconds)
-        end
-        opts
-      end
-
-      # Normalize the spec, in case aliased fields are provided.
-      #
-      # @api private
-      #
-      # @example Normalize the spec.
-      #   Model.normalize_spec(name: 1)
-      #
-      # @param [ Hash ] spec The index specification.
-      #
-      # @return [ Hash ] The normalized specification.
-      #
-      # @since 3.0.7
-      def normalize_spec(spec)
-        spec.inject({}) do |normalized, (name, direction)|
-          normalized[database_field_name(name).to_sym] = direction
-          normalized
-        end
-      end
 
       # Get the names of all databases for this model that have index
       # definitions.
@@ -145,9 +121,23 @@ module Mongoid
       #
       # @since 3.1.0
       def indexed_database_names
-        index_options.values.map do |options|
-          options[:database] || database_name
+        index_specifications.map do |spec|
+          spec.options[:database] || database_name
         end.uniq
+      end
+
+      # Gets a list of index specification keys.
+      #
+      # @api private
+      #
+      # @example Get the specification key list.
+      #   Model.index_keys
+      #
+      # @return [ Array<Hash> ] The specification keys.
+      #
+      # @since 4.0.0
+      def index_keys
+        index_specifications.map{ |spec| spec.key }
       end
     end
   end