+++ spec/mongoid/indexable_spec.rb
@@ -10,12 +10,12 @@ describe Mongoid::Indexable do
       end
     end
 
-    it "adds an index_options accessor" do
-      expect(klass).to respond_to(:index_options)
+    it "adds an index_specifications accessor" do
+      expect(klass).to respond_to(:index_specifications)
     end
 
-    it "defaults index_options to empty hash" do
-      expect(klass.index_options).to eq({})
+    it "defaults index_specifications to empty array" do
+      expect(klass.index_specifications).to be_empty
     end
   end
 
@@ -129,10 +129,12 @@ describe Mongoid::Indexable do
         klass.add_indexes
       end
 
+      let(:spec) do
+        klass.index_specification(_type: 1)
+      end
+
       it "adds the _type index" do
-        expect(klass.index_options[_type: 1]).to eq(
-          { unique: false, background: true }
-        )
+        expect(spec.options).to eq(unique: false, background: true)
       end
     end
   end
@@ -149,11 +151,11 @@ describe Mongoid::Indexable do
     context "when indexing a field that is aliased" do
 
       before do
-        klass.index({ authentication_token: 1 }, { unique: true })
+        klass.index({ authentication_token: 1 }, unique: true)
       end
 
       let(:options) do
-        klass.index_options[a: 1]
+        klass.index_specification(a: 1).options
       end
 
       it "sets the index with unique options" do
@@ -164,11 +166,11 @@ describe Mongoid::Indexable do
     context "when providing unique options" do
 
       before do
-        klass.index({ name: 1 }, { unique: true })
+        klass.index({ name: 1 }, unique: true)
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with unique options" do
@@ -179,11 +181,11 @@ describe Mongoid::Indexable do
     context "when providing a drop_dups option" do
 
       before do
-        klass.index({ name: 1 }, { drop_dups: true })
+        klass.index({ name: 1 }, drop_dups: true)
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with dropDups options" do
@@ -194,11 +196,11 @@ describe Mongoid::Indexable do
     context "when providing a sparse option" do
 
       before do
-        klass.index({ name: 1 }, { sparse: true })
+        klass.index({ name: 1 }, sparse: true)
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with sparse options" do
@@ -209,11 +211,11 @@ describe Mongoid::Indexable do
     context "when providing a name option" do
 
       before do
-        klass.index({ name: 1 }, { name: "index_name" })
+        klass.index({ name: 1 }, name: "index_name")
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with name options" do
@@ -224,11 +226,11 @@ describe Mongoid::Indexable do
     context "when providing database options" do
 
       before do
-        klass.index({ name: 1 }, { database: "mongoid_index_alt" })
+        klass.index({ name: 1 }, database: "mongoid_index_alt")
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with background options" do
@@ -239,11 +241,11 @@ describe Mongoid::Indexable do
     context "when providing a background option" do
 
       before do
-        klass.index({ name: 1 }, { background: true })
+        klass.index({ name: 1 }, background: true)
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with background options" do
@@ -258,11 +260,31 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[name: 1, title: -1]
+        klass.index_specification(name: 1, title: -1).options
       end
 
       it "sets the compound key index" do
         expect(options).to be_empty
+      end
+    end
+
+    context "when providing multiple inverse compound indexes" do
+
+      before do
+        klass.index({ name: 1, title: -1 })
+        klass.index({ title: -1, name: 1 })
+      end
+
+      let(:first_spec) do
+        klass.index_specification(name: 1, title: -1)
+      end
+
+      let(:second_spec) do
+        klass.index_specification(title: -1, name: 1)
+      end
+
+      it "does not overwrite the index options" do
+        expect(first_spec).to_not eq(second_spec)
       end
     end
 
@@ -273,7 +295,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[location: "2d"]
+        klass.index_specification(location: "2d").options
       end
 
       it "sets the geospacial index" do
@@ -288,7 +310,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[location: "geoHaystack"]
+        klass.index_specification(location: "geoHaystack").options
       end
 
       it "sets the geo haystack index" do
@@ -303,7 +325,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[location: "2dsphere"]
+        klass.index_specification(location: "2dsphere").options
       end
 
       it "sets the spherical geospatial index" do
@@ -320,7 +342,7 @@ describe Mongoid::Indexable do
         end
 
         let(:options) do
-          klass.index_options[description: "text"]
+          klass.index_specification(description: "text").options
         end
 
         it "allows the set of the text index" do
@@ -335,7 +357,7 @@ describe Mongoid::Indexable do
         end
 
         let(:options) do
-          klass.index_options[description: "text", name: "text"]
+          klass.index_specification(description: "text", name: "text").options
         end
 
         it "allows the set of the text index" do
@@ -350,7 +372,7 @@ describe Mongoid::Indexable do
         end
 
         let(:options) do
-          klass.index_options[:"$**" => "text"]
+          klass.index_specification(:"$**" => "text").options
         end
 
         it "allows the set of the text index" do
@@ -365,7 +387,7 @@ describe Mongoid::Indexable do
         end
 
         let(:options) do
-          klass.index_options[description: "text"]
+          klass.index_specification(description: "text").options
         end
 
         it "allows the set of the text index" do
@@ -380,7 +402,7 @@ describe Mongoid::Indexable do
         end
 
         let(:options) do
-          klass.index_options[description: "text"]
+          klass.index_specification(description: "text").options
         end
 
         it "allows the set of the text index" do
@@ -396,7 +418,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[a: "hashed"]
+        klass.index_specification(a: "hashed").options
       end
 
       it "sets the hashed index" do
@@ -411,7 +433,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[content: "text"]
+        klass.index_specification(content: "text").options
       end
 
       it "sets the text index" do
@@ -426,7 +448,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[content: "text", title: "text"]
+        klass.index_specification(content: "text", title: "text").options
       end
 
       it "sets the compound text index" do
@@ -441,7 +463,7 @@ describe Mongoid::Indexable do
       end
 
       let(:options) do
-        klass.index_options[name: 1]
+        klass.index_specification(name: 1).options
       end
 
       it "sets the index with sparse options" do